requirements.txt
strands-agents
chromadb
pydantic
requests
python-dotenv
